+++ addon/components/polaris-drop-zone.js
@@ -6,6 +6,7 @@ import { throttle, scheduleOnce } from '@ember/runloop';
 import { isNone, isPresent } from '@ember/utils';
 import ContextBoundEventListenersMixin from 'ember-lifeline/mixins/dom';
 import { getRectForNode } from '@shopify/javascript-utilities/geometry';
+import { guidFor } from '@ember/object/internals';
 import layout from '../templates/components/polaris-drop-zone';
 import State from '../-private/drop-zone-state';
 import {
@@ -20,18 +21,47 @@ const iconDragDrop = 'drag-drop';
 const iconAlertCircle = 'alert-circle';
 
 export default Component.extend(ContextBoundEventListenersMixin, {
-  attributeBindings: ['ariaDisabled:aria-disabled'],
-
-  classNames: ['Polaris-DropZone'],
-
-  classNameBindings: [
-    'outline:Polaris-DropZone--hasOutline',
-    'isDragging:Polaris-DropZone--isDragging',
-    'state.error:Polaris-DropZone--hasError',
-    'sizeClass',
-  ],
-
   layout,
+
+  /**
+   * ID for file input
+   *
+   * @type {String}
+   * @default null
+   * @public
+   * @property id
+   */
+  id: null,
+
+  /**
+   * Label for the file input
+   *
+   * @type {String}
+   * @default null
+   * @public
+   * @property label
+   */
+  label: null,
+
+  /**
+   * Adds an action to the label
+   *
+   * @type {Object}
+   * @default null
+   * @public
+   * @property labelAction
+   */
+  labelAction: null,
+
+  /**
+   * Visually hide the label
+   *
+   * @type {Boolean}
+   * @default false
+   * @public
+   * @property labelHidden
+   */
+  labelHidden: false,
 
   /**
    * Allowed file types
@@ -260,9 +290,41 @@ export default Component.extend(ContextBoundEventListenersMixin, {
 
   isDragging: or('active', 'state.dragging').readOnly(),
 
-  fileInputNode: computed(function() {
+  dropZoneClasses: computed(
+    'outline',
+    'isDragging',
+    'state.error',
+    'sizeClass',
+    function() {
+      let { outline, isDragging, state, sizeClass } = this.getProperties(
+        'outline',
+        'isDragging',
+        'state',
+        'sizeClass'
+      );
+
+      let classNames = ['Polaris-DropZone', sizeClass];
+      let error = state.get('error');
+
+      if (outline) {
+        classNames.push('Polaris-DropZone--hasOutline');
+      }
+
+      if (isDragging) {
+        classNames.push('Polaris-DropZone--isDragging');
+      }
+
+      if (error) {
+        classNames.push('Polaris-DropZone--hasError');
+      }
+
+      return classNames.join(' ');
+    }
+  ),
+
+  fileInputNode: computed('state.id', function() {
     return this.element.querySelector(
-      `input[id='${this.get('elementId')}-input']`
+      `input[id='${this.get('state.id')}-input']`
     );
   }).readOnly(),
 
@@ -285,15 +347,6 @@ export default Component.extend(ContextBoundEventListenersMixin, {
   /**
    * Event handlers
    */
-  handleClick(event) {
-    let { onClick, disabled } = this.getProperties('onClick', 'disabled');
-    if (disabled) {
-      return;
-    }
-
-    return onClick ? onClick(event) : this.open();
-  },
-
   handleDrop(event) {
     event.preventDefault();
     event.stopPropagation();
@@ -410,11 +463,6 @@ export default Component.extend(ContextBoundEventListenersMixin, {
     onDragLeave();
   },
 
-  handleDragStart(event) {
-    event.preventDefault();
-    event.stopPropagation();
-  },
-
   adjustSize() {
     throttle(
       this,
@@ -483,26 +531,30 @@ export default Component.extend(ContextBoundEventListenersMixin, {
     this.addEventListener(dropNode, 'dragenter', this.handleDragEnter);
     this.addEventListener(dropNode, 'dragleave', this.handleDragLeave);
 
-    this.addEventListener(this.element, 'click', this.handleClick);
-    this.addEventListener(this.element, 'dragStart', this.handleDragStart);
-
     this.addEventListener(window, 'resize', this.adjustSize);
   },
 
   setNode() {
     let dropOnPage = this.get('dropOnPage');
+    let dropzoneContainer = this.element.querySelector('.Polaris-DropZone');
+
     this.setProperties({
-      node: this.element,
-      dropNode: dropOnPage ? document : this.element,
+      node: dropzoneContainer,
+      dropNode: dropOnPage ? document : dropzoneContainer,
     });
 
     this.adjustSize();
   },
 
-  updateStateFromProps() {
-    let { error, type, overlayText, errorOverlayText } = this.get(
+  getDerivedStateFromProps() {
+    let { id, error, type, overlayText, errorOverlayText } = this.get(
       'state'
-    ).getProperties('error', 'type', 'overlayText', 'errorOverlayText');
+    ).getProperties('id', 'error', 'type', 'overlayText', 'errorOverlayText');
+
+    let newId = this.get('id') || guidFor(this);
+    if (id !== null && id !== newId) {
+      this.set('state.id', newId);
+    }
 
     if (error !== this.get('error')) {
       this.set('state.error', this.get('error'));
@@ -525,20 +577,26 @@ export default Component.extend(ContextBoundEventListenersMixin, {
     ) {
       this.set('state.errorOverlayText', newErrorOverlayText);
     }
-
-    if (this.get('openFileDialog')) {
-      this.open();
-      scheduleOnce('afterRender', () => this.get('onFileDialogClose')());
-    }
   },
 
   open() {
     let fileInputNode = this.get('fileInputNode');
+
     if (isNone(fileInputNode)) {
       return;
     }
 
     fileInputNode.click();
+  },
+
+  triggerFileDialog() {
+    this.open();
+
+    let close = this.get('onFileDialogClose');
+
+    if (close) {
+      scheduleOnce('afterRender', close);
+    }
   },
 
   /**
@@ -551,7 +609,7 @@ export default Component.extend(ContextBoundEventListenersMixin, {
 
   didReceiveAttrs() {
     this._super(...arguments);
-    this.updateStateFromProps();
+    this.getDerivedStateFromProps();
   },
 
   didInsertElement() {
@@ -560,5 +618,33 @@ export default Component.extend(ContextBoundEventListenersMixin, {
     this.setNode();
     this.set('state.error', this.get('error'));
     this.setupEvents();
+
+    if (this.get('openFileDialog')) {
+      this.triggerFileDialog();
+    }
+  },
+
+  didUpdateAttrs() {
+    this._super(...arguments);
+
+    if (this.get('openFileDialog')) {
+      this.triggerFileDialog();
+    }
+  },
+
+  actions: {
+    handleClick(event) {
+      let { onClick, disabled } = this.getProperties('onClick', 'disabled');
+      if (disabled) {
+        return;
+      }
+
+      return onClick ? onClick(event) : this.open();
+    },
+
+    handleDragStart(event) {
+      event.preventDefault();
+      event.stopPropagation();
+    },
   },
 });